refactor(@formatjs/intl-datetimeformat): split packer.ts into pack (scripts) and unpack (source)
Move pack() to scripts/packer.ts (script-only, uses relative .ts imports)
and extract unpack() into unpack.ts (source, uses #packages/ imports).
This removes packer.ts from the no-relative-import ignore list.
