+++ packages/intl-datetimeformat/scripts/process-zdump.ts
@@ -3,7 +3,7 @@ import {outputFileSync} from 'fs-extra/esm'
 import stringify from 'json-stable-stringify'
 import minimist from 'minimist'
 import {promisify} from 'util'
-import {pack} from '../packer.ts'
+import {pack} from './packer.ts'
 import {type UnpackedData, type ZoneData} from '../types.ts'
 
 const readFile = promisify(_readFile)
